+++ src/style/mod.rs
@@ -1,0 +1,8 @@
+mod style;
+mod styleable;
+mod stylize;
+
+#[allow(clippy::module_inception)]
+pub use style::Style;
+pub use styleable::Styleable;
+pub use stylize::Stylize;